Separate pipeline steps from StatusCard
* Move getDateString, getDateTimeString, and getStage to statusCardUtils
