+++ src/pages/sighting/statusCard/StatusCard.jsx
@@ -1,345 +1,23 @@
 import React from 'react';
-import { useIntl } from 'react-intl';
-import { get } from 'lodash-es';
 
 import Timeline from '@material-ui/lab/Timeline';
-import ReportIcon from '@material-ui/icons/ArtTrack';
-import PreparationIcon from '@material-ui/icons/FileCopy';
-import DetectionIcon from '@material-ui/icons/Search';
-import CurationIcon from '@material-ui/icons/LowPriority';
-import IdentificationIcon from '@material-ui/icons/Visibility';
 
-import {
-  formatDate,
-  formatDateCustom,
-} from '../../../utils/formatters';
-import { getProgress } from '../../../utils/pipelineStatusUtils';
 import Card from '../../../components/cards/Card';
-import ButtonLink from '../../../components/ButtonLink';
-import ProgressMetrics from '../../../components/progress/ProgressMetrics';
-import TimelineStep from './TimelineStep';
-import stages from './stages';
-import useSiteSettings from '../../../models/site/useSiteSettings';
-import wildbookBySystemGuid from '../../../constants/wildbookBySystemGuid';
-
-function getDateString(date) {
-  return date ? formatDate(date, true) : 'unknown date';
-}
-
-function getDateTimeString(date) {
-  return date ? formatDateCustom(date, 'PPp') : '';
-}
-
-function getStage(pipelineStep) {
-  const { skipped, inProgress, complete, failed } =
-    pipelineStep || {};
-
-  if (skipped) return stages.skipped;
-
-  if (inProgress) return stages.current;
-
-  if (failed) return stages.failed;
-
-  if (complete) return stages.finished;
-
-  return stages.waiting;
-}
-
-function scrollToTop() {
-  window.scrollTo(0, 0);
-}
-
-function withNonWrappingSpan(chunk) {
-  return <span style={{ whiteSpace: 'nowrap' }}>{chunk}</span>;
-}
+import SubmissionStep from './SubmissionStep';
+import PreparationStep from './PreparationStep';
+import DetectionStep from './DetectionStep';
+import CurationStep from './CurationStep';
+import IdentificationStep from './IdentificationStep';
 
 export default function StatusCard({ sightingData }) {
-  const intl = useIntl();
-  const systemGuid = get(useSiteSettings(), 'data.system_guid.value');
-  const migratedSiteName = get(wildbookBySystemGuid, systemGuid);
-
-  const assets = get(sightingData, 'assets', []);
-  const assetCount = assets.length;
-  const dateCreated = get(sightingData, 'submissionTime');
-  const currentUserHasEditPermission = get(
-    sightingData,
-    'hasEdit',
-    false,
-  );
-
-  const {
-    preparation: preparationStep,
-    detection: detectionStep,
-    curation: curationStep,
-    identification: identificationStep,
-    migrated,
-  } = sightingData?.pipeline_status || {};
-
-  const {
-    start: preparationStartTime,
-    end: preparationEndTime,
-    inProgress: isPreparationInProgress,
-    message: preparationMessage,
-  } = preparationStep || {};
-
-  const preparationStage = getStage(preparationStep);
-
-  const formattedPreparationStartTime = getDateTimeString(
-    preparationStartTime,
-  );
-
-  const {
-    start: detectionStartTime,
-    end: detectionEndTime,
-    inProgress: isDetectionInProgress,
-    message: detectionMessage,
-  } = detectionStep || {};
-
-  const detectionStage = getStage(detectionStep);
-  let detectionSkippedLabelId = 'DETECTION_SKIPPED_MESSAGE';
-  if (assetCount === 0) {
-    detectionSkippedLabelId = 'DETECTION_SKIPPED_NO_IMAGES_MESSAGE';
-  } else if (
-    get(sightingData, 'speciesDetectionModel[0]') === 'None'
-  ) {
-    detectionSkippedLabelId = 'DETECTION_SKIPPED_NO_MODEL_MESSAGE';
-  }
-
-  const formattedDetectionStartTime = getDateTimeString(
-    detectionStartTime,
-  );
-
-  const {
-    end: curationEndTime,
-    inProgress: isCurationInProgress,
-    message: curationMessage,
-  } = curationStep || {};
-
-  const curationStage = getStage(curationStep);
-  const someAssetsHaveAnnotations = assets.some(
-    asset => get(asset, 'annotations.length', 0) > 0,
-  );
-
-  let curationFinishedLabel;
-  if (migrated) {
-    curationFinishedLabel = migratedSiteName
-      ? intl.formatMessage(
-          { id: 'MIGRATION_FINISHED_MESSAGE_SITE' },
-          { date: getDateString(curationEndTime), migratedSiteName },
-        )
-      : intl.formatMessage(
-          { id: 'MIGRATION_FINISHED_MESSAGE_DEFAULT' },
-          { date: getDateString(curationEndTime) },
-        );
-  } else {
-    curationFinishedLabel = intl.formatMessage(
-      { id: 'CURATION_FINISHED_MESSAGE' },
-      { date: getDateString(curationEndTime) },
-    );
-  }
-
-  const {
-    start: identificationStartTime,
-    end: identificationEndTime,
-    inProgress: isIdentificationInProgress,
-    complete: isIdentificationComplete,
-    message: identificationMessage,
-  } = identificationStep || {};
-
-  const identificationStage = getStage(identificationStep);
-  let identificationSkippedLabel;
-  if (migrated) {
-    identificationSkippedLabel = migratedSiteName
-      ? intl.formatMessage(
-          { id: 'IDENTIFICATION_SKIPPED_MIGRATED_MESSAGE_SITE' },
-          { migratedSiteName },
-        )
-      : intl.formatMessage({
-          id: 'IDENTIFICATION_SKIPPED_MIGRATED_MESSAGE_DEFAULT',
-        });
-  } else if (assetCount === 0) {
-    identificationSkippedLabel = intl.formatMessage({
-      id: 'IDENTIFICATION_SKIPPED_NO_IMAGES_MESSAGE',
-    });
-  } else {
-    identificationSkippedLabel = intl.formatMessage({
-      id: 'IDENTIFICATION_SKIPPED_MESSAGE',
-    });
-  }
-
-  const formattedIdentificationStartTime = getDateTimeString(
-    identificationStartTime,
-  );
-
   return (
     <Card titleId="IDENTIFICATION_PIPELINE_STATUS" maxHeight={900}>
       <Timeline>
-        <TimelineStep
-          Icon={ReportIcon}
-          titleId="SIGHTING_SUBMISSION"
-          stage={stages.finished}
-          finishedText={intl.formatMessage(
-            { id: 'SIGHTING_SUBMISSION_REPORT_DATE' },
-            { date: getDateString(dateCreated) },
-          )}
-        />
-        <TimelineStep
-          Icon={PreparationIcon}
-          titleId="SIGHTING_PREPARATION"
-          stage={preparationStage}
-          notStartedText={intl.formatMessage({
-            id: 'WAITING_ELLIPSES',
-          })}
-          inProgressText={
-            formattedPreparationStartTime
-              ? intl.formatMessage(
-                  { id: 'SIGHTING_PREPARATION_STARTED_ON_MESSAGE' },
-                  { date: formattedPreparationStartTime },
-                )
-              : intl.formatMessage({
-                  id: 'SIGHTING_PREPARATION_STARTED_ON_UNKNOWN_MESSAGE',
-                })
-          }
-          finishedText={intl.formatMessage(
-            { id: 'SIGHTING_PREPARATION_FINISHED_MESSAGE' },
-            {
-              photoCount: assetCount,
-              date: getDateString(preparationEndTime),
-              nonWrapping: withNonWrappingSpan,
-            },
-          )}
-          skippedText={intl.formatMessage({
-            id: 'SIGHTING_PREPARATION_SKIPPED_MESSAGE',
-          })}
-          failedText={intl.formatMessage({
-            id: 'SIGHTING_PREPARATION_FAILED',
-          })}
-          failedAlertDescription={preparationMessage}
-        >
-          {isPreparationInProgress && (
-            <ProgressMetrics
-              progress={getProgress(preparationStep)}
-              style={{ marginTop: 20 }}
-            />
-          )}
-        </TimelineStep>
-        <TimelineStep
-          Icon={DetectionIcon}
-          titleId="ANIMAL_DETECTION"
-          stage={detectionStage}
-          notStartedText={intl.formatMessage({
-            id: 'WAITING_ELLIPSES',
-          })}
-          inProgressText={
-            formattedDetectionStartTime
-              ? intl.formatMessage(
-                  { id: 'DETECTION_STARTED_ON_MESSAGE' },
-                  { date: formattedDetectionStartTime },
-                )
-              : intl.formatMessage({
-                  id: 'DETECTION_STARTED_ON_UNKNOWN_MESSAGE',
-                })
-          }
-          finishedText={intl.formatMessage(
-            { id: 'DETECTION_FINISHED_MESSAGE' },
-            { date: getDateString(detectionEndTime) },
-          )}
-          skippedText={intl.formatMessage({
-            id: detectionSkippedLabelId,
-          })}
-          failedText={intl.formatMessage({ id: 'DETECTION_FAILED' })}
-          failedAlertDescription={detectionMessage}
-        >
-          {isDetectionInProgress && (
-            <ProgressMetrics
-              progress={getProgress(detectionStep)}
-              style={{ marginTop: 20 }}
-            />
-          )}
-        </TimelineStep>
-        <TimelineStep
-          Icon={CurationIcon}
-          titleId="SIGHTING_CURATION"
-          stage={curationStage}
-          notStartedText={intl.formatMessage({
-            id: 'WAITING_ELLIPSES',
-          })}
-          inProgressText={
-            currentUserHasEditPermission
-              ? intl.formatMessage({ id: 'CURATION_INSTRUCTIONS' })
-              : intl.formatMessage({ id: 'IN_PROGRESS' })
-          }
-          finishedText={curationFinishedLabel}
-          skippedText={intl.formatMessage({
-            id: 'CURATION_SKIPPED_MESSAGE',
-          })}
-          failedText={intl.formatMessage({ id: 'CURATION_FAILED' })}
-          failedAlertDescription={curationMessage}
-        >
-          {isCurationInProgress && currentUserHasEditPermission && (
-            <div style={{ marginTop: 4, marginBottom: 20 }}>
-              <ButtonLink
-                id={
-                  someAssetsHaveAnnotations
-                    ? 'CURATION_ASSIGN_ANNOTATIONS'
-                    : 'CURATION_ANNOTATE_PHOTOS'
-                }
-                href={
-                  someAssetsHaveAnnotations
-                    ? '#individuals'
-                    : '#photographs'
-                }
-                display="primary"
-                size="small"
-                onClick={scrollToTop}
-              />
-            </div>
-          )}
-        </TimelineStep>
-        <TimelineStep
-          Icon={IdentificationIcon}
-          titleId="IDENTIFICATION"
-          stage={identificationStage}
-          notStartedText={intl.formatMessage({
-            id: 'WAITING_ELLIPSES',
-          })}
-          inProgressText={
-            formattedIdentificationStartTime
-              ? intl.formatMessage(
-                  { id: 'IDENTIFICATION_STARTED_ON_MESSAGE' },
-                  { date: formattedIdentificationStartTime },
-                )
-              : intl.formatMessage({
-                  id: 'IDENTIFICATION_STARTED_ON_UNKNOWN_MESSAGE',
-                })
-          }
-          finishedText={intl.formatMessage(
-            { id: 'IDENTIFICATION_FINISHED_MESSAGE' },
-            { date: getDateString(identificationEndTime) },
-          )}
-          skippedText={identificationSkippedLabel}
-          failedText={intl.formatMessage({
-            id: 'IDENTIFICATION_FAILED',
-          })}
-          failedAlertDescription={identificationMessage}
-        >
-          {isIdentificationInProgress && (
-            <ProgressMetrics
-              progress={getProgress(identificationStep)}
-              style={{ marginTop: 20 }}
-            />
-          )}
-          {isIdentificationComplete && (
-            <div style={{ marginTop: 4 }}>
-              <ButtonLink
-                id="IDENTIFICATION_VIEW_RESULTS"
-                href={`/match-results/${sightingData?.guid}`}
-                display="panel"
-                size="small"
-              />
-            </div>
-          )}
-        </TimelineStep>
+        <SubmissionStep sightingData={sightingData} />
+        <PreparationStep sightingData={sightingData} />
+        <DetectionStep sightingData={sightingData} />
+        <CurationStep sightingData={sightingData} />
+        <IdentificationStep sightingData={sightingData} />
       </Timeline>
     </Card>
   );
